backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
pandas==2.1.3
numpy==1.25.2
talib-binary==0.4.28
python-dotenv==1.0.0
websockets==12.0
asyncio==3.4.3
